Adds edge-based node contractor and PrepareCH classes.
* Adds the possibility to automatically find shortcuts and contract a
given graph in an edge-based fashion.

* This enables turn costs & restrictions in contraction hierarchies.

* EdgeBasedNodeContractor and EdgeBasedPrepareContractionHierarchies
are not yet integrated with the previous node based CH code and so far
no automatic node ordering can be calculated, but for correctness it is
most important that the query results are correct for ANY contraction order.
For automatic node contraction ordering EdgeBasedNodeContractor needs
to return the required metrics (to calculate edge difference / quotient etc.).

* This commit only focuses on correctness and adds extensive tests,
there are many border cases with parallel shortcuts, loops etc. that
have to be accounted for.

* No support for U-Turn traversal mode (not sure if it even makes sense)

* Shortcut edges are used in fwd direction only, so the memory consumption
is potentially (almost) doubled (assuming that most shortcuts can be travelled
in both directions), but using fwd&bwd flags requires some changes because
then it is harder to decide what the 'first' and 'last' original edges are.

* There are some very obvious potential performance improvements that
are left to be done, some are mentioned as todos. So far the contraction
has not been tested on a bigger graph.

Signed-off-by: ammagamma <contactammagamma@gmail.com>